Improved performance
It may not seem significant, however, if you have a lot of text to process it will be very slow. That regex pattern is also very lengthy and takes up a lot of space. My method enumerates over the words and checks, if it is in a stop_words set which significantly speeds it up. It performs same operation as the regex, however it is much faster. Around 20x faster.